--- src/MakingSense.AspNet.HypermediaApi.Seed/project.json
@@ -1,37 +1,0 @@
-﻿{
-  "webroot": "wwwroot",
-  "version": "1.0.0-*",
-
-  "dependencies": {
-    "MakingSense.AspNet.Authentication.SimpleToken": "0.1.0-rc1-final-14",
-    "MakingSense.AspNet.Documentation": "0.1.0-rc1-final-13",
-    "MakingSense.AspNet.HypermediaApi": "0.1.0-rc1-final-36",
-    "Microsoft.AspNet.IISPlatformHandler": "1.0.0-rc1-final",
-    "Microsoft.AspNet.Mvc": "6.0.0-rc1-final",
-    "Microsoft.AspNet.Server.Kestrel": "1.0.0-rc1-final",
-    "Microsoft.AspNet.Server.WebListener": "1.0.0-rc1-final",
-    "Microsoft.AspNet.StaticFiles": "1.0.0-rc1-final"
-  },
-
-  "commands": {
-    "web": "Microsoft.AspNet.Server.Kestrel"
-  },
-
-  "frameworks": {
-    "dnx451": { },
-    "dnxcore50": { }
-  },
-
-  "publishExclude": [
-    "node_modules",
-    "bower_components",
-    "**.xproj",
-    "**.user",
-    "**.vspscc"
-  ],
-  "exclude": [
-    "wwwroot",
-    "node_modules",
-    "bower_components"
-  ]
-}
