--- src/MakingSense.AspNet.HypermediaApi.Seed/wwwroot/web.config
@@ -1,24 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <system.webServer>
-    <handlers>
-      <clear />
-      <add name="httpPlatformHandler" path="*" verb="*" modules="httpPlatformHandler" resourceType="Unspecified" />
-    </handlers>
-    <security>
-      <requestFiltering allowDoubleEscaping="true">
-        <fileExtensions>
-          <clear />
-        </fileExtensions>
-        <verbs>
-          <clear />
-          <add verb="DELETE" allowed="true" />
-          <add verb="POST" allowed="true" />
-          <add verb="PUT" allowed="true" />
-          <add verb="GET" allowed="true" />
-        </verbs>
-      </requestFiltering>
-    </security>
-    <httpPlatform processPath="%DNX_PATH%" arguments="%DNX_ARGS%" forwardWindowsAuthToken="false" startupTimeLimit="3600" />
-  </system.webServer>
-</configuration>
