+++ src/MakingSense.AspNetCore.HypermediaApi.Seed/DocumentationTemplates/Layout.html
