+++ src/MakingSense.AspNetCore.HypermediaApi.Seed/DocumentationTemplates/NotFound.html
