+++ src/MakingSense.AspNetCore.HypermediaApi.Seed/Properties/launchSettings.json
