+++ src/MakingSense.AspNetCore.HypermediaApi.Seed/wwwroot/docs/index.md
