+++ src/MakingSense.AspNetCore.HypermediaApi.Seed/wwwroot/docs/models/message.md
