+++ src/MakingSense.AspNetCore.HypermediaApi.Seed/wwwroot/docs/rels/get-account-home.md
