+++ src/MakingSense.AspNetCore.HypermediaApi.Seed/wwwroot/docs/rels/get-index.md
