+++ src/MakingSense.AspNetCore.HypermediaApi.Seed/wwwroot/schemas/link-collection.json
