+++ src/MakingSense.AspNetCore.HypermediaApi.Seed/wwwroot/schemas/message.json
