+++ src/MakingSense.AspNetCore.HypermediaApi.Seed/wwwroot/schemas/problem.json
