+++ src/MakingSense.AspNetCore.HypermediaApi.Seed/wwwroot/web.config
@@ -1,0 +1,24 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+    <system.webServer>
+        <handlers>
+            <clear />
+            <add name="aspNetCore" path="*" verb="*" modules="AspNetCoreModule" resourceType="Unspecified"/>
+        </handlers>
+        <security>
+            <requestFiltering allowDoubleEscaping="true">
+                <fileExtensions>
+                    <clear />
+                </fileExtensions>
+                <verbs>
+                    <clear />
+                    <add verb="DELETE" allowed="true" />
+                    <add verb="POST" allowed="true" />
+                    <add verb="PUT" allowed="true" />
+                    <add verb="GET" allowed="true" />
+                </verbs>
+            </requestFiltering>
+        </security>
+        <aspNetCore processPath="%LAUNCHER_PATH%" arguments="%LAUNCHER_ARGS%" stdoutLogEnabled="false" stdoutLogFile=".\logs\stdout" forwardWindowsAuthToken="false" />
+    </system.webServer>
+</configuration>
